requirements.txt
pprint
pretrainedmodels
pytorch
tqdm